Move output traits and types to `output` sub module.
