+++ crate/rt/src/cmds/states_saved_display_cmd.rs
@@ -5,7 +5,7 @@ use peace_resources::{
     Resources,
 };
 use peace_rt_model::{CmdContext, Error};
-use peace_rt_model_core::OutputWrite;
+use peace_rt_model_core::output::OutputWrite;
 
 use crate::cmds::sub::StatesSavedReadCmd;
