+++ workspace_tests/src/fn_tracker_output.rs
@@ -4,7 +4,7 @@ use peace::{
         StateDiffs, StatesCleaned, StatesCleanedDry, StatesDesired, StatesEnsured,
         StatesEnsuredDry, StatesSaved,
     },
-    rt_model::OutputWrite,
+    rt_model::output::OutputWrite,
 };
 
 use crate::FnInvocation;
