+++ workspace_tests/src/rt_model.rs
@@ -1,15 +1,8 @@
-mod cli_output;
 mod cmd_context;
 mod cmd_context_builder;
 mod item_spec_boxed;
 mod item_spec_wrapper;
 mod outcomes;
-mod output_format;
-mod output_format_parse_error;
+mod output;
 mod states_serializer;
 mod workspace_dirs_builder;
-
-#[cfg(feature = "output_progress")]
-mod cli_progress_format;
-#[cfg(feature = "output_progress")]
-mod cli_progress_format_parse_error;
