+++ workspace_tests/src/rt_model/output.rs
@@ -1,0 +1,7 @@
+mod cli_output;
+#[cfg(feature = "output_progress")]
+mod cli_progress_format;
+#[cfg(feature = "output_progress")]
+mod cli_progress_format_parse_error;
+mod output_format;
+mod output_format_parse_error;
