+++ workspace_tests/src/rt_model/output/cli_progress_format_parse_error.rs
@@ -1,4 +1,4 @@
-use peace::rt_model::CliProgressFormatParseError;
+use peace::rt_model::output::CliProgressFormatParseError;
 
 #[test]
 fn display_includes_auto_output_pb_progress_bar() {
